Add Cloudinary integration for image uploads and update leave management

+++ src/main/java/org/ems/demo/service/LeaveService.java
@@ -17,9 +17,9 @@ public interface LeaveService {
     void deleteLeave(Long id);
 
     List<Leave> getAllLeavesByUser(Integer userId, int limit, int offset);
-    Map<String, Integer> getLeaveCounts(Long companyId);
+    Map<String, Double> getLeaveCounts(Long companyId);
 
     Integer getLeaveCountsByUser(Integer userId);
 
-    Map<String,Integer> getLeaveCountsDatesByUser(Integer userId);
+    Map<String,Double> getLeaveCountsDatesByUser(Integer userId);
 }
